Reserva con pago todo en uno, renderiza un componente aparte, no pude hacer popup
